Fix script for ssh server selection

+++ scripts/ssh.sh
@@ -4,9 +4,17 @@ MYDIR=`dirname $0`
 . $MYDIR/include.sh
 
 if [ -z "$@" ]; then
-       exec "$TERMINAL" "$CLASS_OPT" "xmDialog" -e $0 x
+	TEMPFILE=`mktemp -t xmonad-script-output`
+	"$TERMINAL" "$CLASS_OPT" "xmDialog" -e $0 "$TEMPFILE"
+	EXEC=`cat $TEMPFILE`
+	rm -f "$TEMPFILE"
+	if [ -z "$EXEC" ]; then
+		exit 0
+	fi
+	exec sh -c "$EXEC"
 fi
 
+TEMPFILE="$1"
 CACHE_FILE="$CACHE_DIR/ssh-last.txt"
 
 DEFAULT=`test -r "$CACHE_FILE" && cat "$CACHE_FILE" | awk '!seen[$0]++' | tail -n 10`
@@ -18,28 +26,29 @@ for l in $DEFAULT; do
 	SELECTED="off"
 done
 
-TEMPFILE=`mktemp -t xmonad-script-output`
 CANCELLED=0
-dialog --no-items --radiolist "SSH connection:" 50 70 45 $LOGINS 2> "$TEMPFILE" || CANCELLED=1
-sleep 2
+dialog --no-items --radiolist "SSH connection:" 50 70 45 $LOGINS "Add a new entry" off 2> "$TEMPFILE" || CANCELLED=1
 INPUT=`cat "$TEMPFILE"`
 
 if [ $CANCELLED -ne 0 ]; then
+	exit 1
+fi
+
+if [ "$INPUT" = "Add a new entry" ]; then
 	CANCELLED=0
 	dialog --inputbox "New SSH connection:\n[user@]hostname[:port]?" 20 40 2> "$TEMPFILE" || CANCELLED=1
 	INPUT=`cat "$TEMPFILE"`
 	if [ $CANCELLED -ne 0 ]; then
 		rm -f "$TEMPFILE"
-		exit 0
+		exit 1
 	fi
 fi
-rm -f "$TEMPFILE"
 
 SSH_LOGIN=`echo "$INPUT" | sed 's/:[^:]*$//'`
-SSH_OPT=`echo "$INPUT" | sed 's/^.*://; /^$/q; s/^/-p /'`
+SSH_OPT=`echo "$INPUT" | grep ':' | sed 's/^.*://; /^$/q; s/^/-p /'`
 
 if [ -z "$SSH_LOGIN" ]; then
-	exit 0
+	exit 1
 fi
 
 mkdir -p "$CACHE_DIR" && ( echo "$INPUT" > "$CACHE_FILE" ;
@@ -49,4 +58,5 @@ mkdir -p "$CACHE_DIR" && ( echo "$INPUT" > "$CACHE_FILE" ;
 			fi
 		done )
 
-ssh $SSH_OPT "$SSH_LOGIN"
+echo "$DEFAULT_X_TERMINAL -e ssh -t $SSH_OPT '$SSH_LOGIN' 'tmux -2 new-session'" > "$TEMPFILE"
+exit 0
